Use macros to create versions of Frodo functions with hard-coded parameters to enable compiler optimization
